+++ gallery_dl/extractor/twitter.py
@@ -546,15 +546,17 @@ class TwitterTimelineExtractor(TwitterExtractor):
     def _select_tweet_source(self):
         strategy = self.config("strategy")
         if strategy is None or strategy == "auto":
-            if self.retweets or self.replies or self.textonly:
+            if self.retweets or self.textonly:
                 return self.api.user_tweets
             else:
                 return self.api.user_media
         if strategy == "tweets":
             return self.api.user_tweets
+        if strategy == "media":
+            return self.api.user_media
         if strategy == "with_replies":
             return self.api.user_tweets_and_replies
-        return self.api.user_media
+        raise exception.StopExtraction("Invalid strategy '%s'", strategy)
 
 
 class TwitterTweetsExtractor(TwitterExtractor):